# enh: add disableClientMask option for WebSocket payload masking and optimize mask calculation