<!--
     For Work In Progress Pull Requests, please use the Draft PR feature,
     see https://github.blog/2019-02-14-introducing-draft-pull-requests/ for further details.

     For a timely review/response, please avoid force-pushing additional
     commits if your PR already received reviews or comments.

     Before submitting a Pull Request, please ensure that you have:
     - 📖 Read the Contributing guide: https://github.com/gorilla/.github/blob/main/CONTRIBUTING.md
     - 📖 Read the Code of Conduct: https://github.com/gorilla/.github/blob/main/CODE_OF_CONDUCT.md

     - Provide tests for your changes.
     - Use descriptive commit messages.
	 - Comment your code where appropriate.
	 - Squash your commits
     - Update any related documentation.

     - Add gorilla/pull-request-reviewers as a Reviewer
-->

## What type of PR is this? (check all applicable)

- [ ] Refactor
- [ ] Feature
- [ ] Bug Fix
- [ ] Optimization
- [ ] Documentation Update
- [ ] Go Version Update
- [ ] Dependency Update

## Description
Optimized WebSocket masking by:
1. Implementing fast-path bypass when mask key equals [4]byte{0,0,0,0}
2. Adding configurable masking disablement for secure contexts (e.g., TLS)
```text
goos: linux
goarch: amd64
pkg: github.com/gorilla/websocket
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMaskBytes
BenchmarkMaskBytes/size-2
BenchmarkMaskBytes/size-2/align-4
BenchmarkMaskBytes/size-2/align-4/byte
BenchmarkMaskBytes/size-2/align-4/byte-16         	455632444	         2.629 ns/op	 760.76 MB/s
BenchmarkMaskBytes/size-2/align-4/wordV1
BenchmarkMaskBytes/size-2/align-4/wordV1-16       	421480846	         2.849 ns/op	 702.07 MB/s
BenchmarkMaskBytes/size-2/align-4/word
BenchmarkMaskBytes/size-2/align-4/word-16         	423765326	         2.828 ns/op	 707.18 MB/s
BenchmarkMaskBytes/size-4
BenchmarkMaskBytes/size-4/align-4
BenchmarkMaskBytes/size-4/align-4/byte
BenchmarkMaskBytes/size-4/align-4/byte-16         	290015848	         4.140 ns/op	 966.14 MB/s
BenchmarkMaskBytes/size-4/align-4/wordV1
BenchmarkMaskBytes/size-4/align-4/wordV1-16       	323833113	         3.735 ns/op	1071.01 MB/s
BenchmarkMaskBytes/size-4/align-4/word
BenchmarkMaskBytes/size-4/align-4/word-16         	323727016	         3.692 ns/op	1083.36 MB/s
BenchmarkMaskBytes/size-8
BenchmarkMaskBytes/size-8/align-4
BenchmarkMaskBytes/size-8/align-4/byte
BenchmarkMaskBytes/size-8/align-4/byte-16         	220623301	         5.432 ns/op	1472.64 MB/s
BenchmarkMaskBytes/size-8/align-4/wordV1
BenchmarkMaskBytes/size-8/align-4/wordV1-16       	212184852	         5.671 ns/op	1410.73 MB/s
BenchmarkMaskBytes/size-8/align-4/word
BenchmarkMaskBytes/size-8/align-4/word-16         	211640461	         5.642 ns/op	1417.93 MB/s
BenchmarkMaskBytes/size-16
BenchmarkMaskBytes/size-16/align-4
BenchmarkMaskBytes/size-16/align-4/byte
BenchmarkMaskBytes/size-16/align-4/byte-16        	128340018	         9.387 ns/op	1704.47 MB/s
BenchmarkMaskBytes/size-16/align-4/wordV1
BenchmarkMaskBytes/size-16/align-4/wordV1-16      	85413890	        14.01 ns/op	1142.38 MB/s
BenchmarkMaskBytes/size-16/align-4/word
BenchmarkMaskBytes/size-16/align-4/word-16        	125297701	         9.564 ns/op	1673.03 MB/s
BenchmarkMaskBytes/size-32
BenchmarkMaskBytes/size-32/align-4
BenchmarkMaskBytes/size-32/align-4/byte
BenchmarkMaskBytes/size-32/align-4/byte-16        	70067914	        17.31 ns/op	1848.34 MB/s
BenchmarkMaskBytes/size-32/align-4/wordV1
BenchmarkMaskBytes/size-32/align-4/wordV1-16      	81590757	        14.68 ns/op	2180.16 MB/s
BenchmarkMaskBytes/size-32/align-4/word
BenchmarkMaskBytes/size-32/align-4/word-16        	90093694	        13.32 ns/op	2402.45 MB/s
BenchmarkMaskBytes/size-512
BenchmarkMaskBytes/size-512/align-4
BenchmarkMaskBytes/size-512/align-4/byte
BenchmarkMaskBytes/size-512/align-4/byte-16       	 4654480	       256.3 ns/op	1997.71 MB/s
BenchmarkMaskBytes/size-512/align-4/wordV1
BenchmarkMaskBytes/size-512/align-4/wordV1-16     	30228372	        39.81 ns/op	12862.16 MB/s
BenchmarkMaskBytes/size-512/align-4/word
BenchmarkMaskBytes/size-512/align-4/word-16       	38939992	        30.81 ns/op	16616.76 MB/s
BenchmarkMaskBytes/size-1024
BenchmarkMaskBytes/size-1024/align-4
BenchmarkMaskBytes/size-1024/align-4/byte
BenchmarkMaskBytes/size-1024/align-4/byte-16      	 2361459	       506.7 ns/op	2020.90 MB/s
BenchmarkMaskBytes/size-1024/align-4/wordV1
BenchmarkMaskBytes/size-1024/align-4/wordV1-16    	18674653	        64.18 ns/op	15955.39 MB/s
BenchmarkMaskBytes/size-1024/align-4/word
BenchmarkMaskBytes/size-1024/align-4/word-16      	22626250	        52.46 ns/op	19518.92 MB/s
BenchmarkMaskBytes/size-1048576
BenchmarkMaskBytes/size-1048576/align-4
BenchmarkMaskBytes/size-1048576/align-4/byte
BenchmarkMaskBytes/size-1048576/align-4/byte-16   	    2314	    517830 ns/op	2024.94 MB/s
BenchmarkMaskBytes/size-1048576/align-4/wordV1
BenchmarkMaskBytes/size-1048576/align-4/wordV1-16 	   23004	     53214 ns/op	19704.72 MB/s
BenchmarkMaskBytes/size-1048576/align-4/word
BenchmarkMaskBytes/size-1048576/align-4/word-16   	   25855	     46638 ns/op	22483.08 MB/s
```
## Related Tickets & Documents

<!--
For pull requests that relate or close an issue, please include them
below.  We like to follow [Github's guidance on linking issues to pull requests](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue).

For example having the text: "closes #1234" would connect the current pull
request to issue 1234.  And when we merge the pull request, Github will
automatically close the issue.
-->

- Related Issue #
- Closes #

## Added/updated tests?

- [ ] Yes
- [ ] No, and this is why: _please replace this line with details on why tests
      have not been included_
- [ ] I need help with writing tests

## Run verifications and test

- [ ] `make verify` is passing
- [ ] `make test` is passing